quantum-sol-stack/requirements.txt
# Core
pandas>=1.5.0
numpy>=1.21.0

# Data Sources
requests>=2.28.0
python-dotenv>=0.19.0

# Visualization
plotly>=5.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Backtesting
pytest>=7.0.0

# Utils
tqdm>=4.62.0
python-dateutil>=2.8.2
pytz>=2021.3
